# perf: Skip a write to storage when nothing is actually changing