@kidroca 2021-07-29
Note longer total time compared to previous PR due to **preferredLocale**
Also a bigger report is used as my app started to init on a different report and I couldn't remember the report used on prior PRs
  - see: https://expensify.slack.com/archives/C01GTK53T8Q/p1627310018264000
  
This PR together with the `preferedLocale` fixes should bring Total Time spent in Android to ~60sec. 
`preferedLocale` takes about 40-50 on the report used in the benchmark