@quinthar 2021-07-29
nice!!

On Thu, Jul 29, 2021 at 12:27 PM Peter Velkov ***@***.***>
wrote:

> Note longer total time compared to previous PR due to *preferredLocale*
> Also a bigger report is used as my app started to init on a different
> report and I couldn't remember the report used on prior PRs
>
>    - see:
>    https://expensify.slack.com/archives/C01GTK53T8Q/p1627310018264000
>
> This PR together with the preferedLocale fixes should bring Total Time
> spent in Android to ~60sec.
> preferedLocale takes about 40-50 on the report used in the benchmark
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/Expensify/react-native-onyx/pull/92#issuecomment-889400843>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEMNUXLEUCUCZX4ONKFFA3T2GTRXANCNFSM5BHDBZBQ>
> .
>